+++ middleware.ts
@@ -1,5 +1,5 @@
 import { NextResponse, type NextRequest } from 'next/server'
-import { createClient } from '@/auth/supabase/middleware'
+import { createClient } from '@/(auth)/supabase/middleware'
 
 export async function middleware(request: NextRequest) {
   const { supabase, response } = createClient(request)
@@ -11,13 +11,13 @@ export async function middleware(request: NextRequest) {
     data: { session }
   } = await supabase.auth.getSession()
 
-  const loginUri = '/auth/login'
+  const signinUri = '/signin'
 
-  if (request.nextUrl.pathname.startsWith(loginUri)) {
+  if (request.nextUrl.pathname.startsWith(signinUri)) {
     return response
   }
   if (error || !session) {
-    return NextResponse.redirect(new URL(loginUri, request.url))
+    return NextResponse.redirect(new URL(signinUri, request.url))
   }
 
   return response
